# SourceCode/swagger.yml
...
openapi: 3.0.0
info:
  title: "Trail microservice API"
  description: "An API for trails"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
...
    TrailFeature:
      type: "object"
      properties:
        FeatureId:
          type: "integer"
        TrailId:
          type: "integer"
    TrailLocationPoint:
      type: "object"
      properties:
        LocationPointId:
          type: "integer"
        TrailId:
          type: "integer"
        OrderNum:
          type: "integer"
  parameters:
    LocationPointId:
      name: "LocationPointId"
      description: "Id of a location point"
      in: path
      required: True
      schema:
        type: "integer"
    FeatureId:
      name: "FeatureId"
      description: "Id of the feature"
      in: path
      required: True
      schema:
        type: "integer"
    TrailId:
      name: "TrailId"
      description: "Id of a trail"
      in: path
      required: True
      schema:
        type: "integer"
...
paths:
...
  /trails/{TrailId}:

    get:
      operationId: "trails.getTrailById"
      tags:
        - Trails
      summary: "Get trail by Id"
      parameters:
       - $ref: "#/components/parameters/TrailId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "No Id provided"
        "404":
          description: "Trail not found"
        "500":
          description: "Server error"

    delete:
      operationId: "trails.deleteTrailById"
      tags:
        - Trails
      summary: "Delete trail by Id"
      parameters:
       - $ref: "#/components/parameters/TrailId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Trail not found"
        "500":
          description: "Server error"
...
  /features/{featureId}:

    get:
      operationId: "features.getFeaturePointById"
      tags:
        - Features
      summary: "Gets feature with given Id"
      parameters:
       - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Feature not found"
        "500":
          description: "Server error"


    delete:
      operationId: "features.deleteFeatureById"
      tags:
        - Features
      summary: "Deletes feature with given id"
      parameters:
       - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Feature not found"
        "500":
          description: "Server error"

# Features
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# Location Points
...
  /locationpoints/{LocationPointId}:
    get:
      operationId: "locationpoint.getLocationPointById"
      tags:
        - Location points
      summary: "Gets location point with given Id"
      parameters:
       - $ref: "#/components/parameters/LocationPointId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Location point not found"
        "500":
          description: "Server error"

    delete:
      operationId: "locationpoints.deleteLocationPointById"
      tags:
        - Location points
      summary: "Deletes location point with given id"
      parameters:
       - $ref: "#/components/parameters/LocationPointId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Location point not found"
        "500":
          description: "Server error"

# Location Points
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# Link entities
  /traillocationlink:
    post:
      operationId: "linkhelper.newTrailLocationPoint"
      tags:
        - Link entity
      summary: "Add new trail-location point link"
      requestBody:
        description: "Location point and trail to link"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailLocationPointJSON"
              $ref: "#/components/schemas/TrailLocationPoint"
      responses:
        "201":
          description: "Successfully added trail location point"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"

    delete:
      operationId: "linkhelper.deleteTrailLocationPoint"
      tags:
        - Link entity
      summary: "Delete trail-location point link"
      requestBody:
        description: "Location point and trail to delete link of"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailLocationPointJSON"
              $ref: "#/components/schemas/TrailLocationPoint"
      responses:
        "201":
          description: "Successfully deleted trail location point"
        "400":
          description: "Bad request"
        "404":
          description: "No TrailLocationPoint found"
        "500":
          description: "Server error"
  /traillocationlink/gettrails/{LocationPointId}:
    get:
      operationId: "linkhelper.getTrailsByLocationPoint"
      tags:
        - Link entity
      summary: "Get all trails with given location point"
      parameters:
        - $ref: "#/components/parameters/LocationPointId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "None found"
        "500":
          description: "Server error"
  /traillocationlink/getpoints/{TrailId}:
    get:
      operationId: "linkhelper.getLocationPointsByTrail"
      tags:
        - Link entity
      summary: "Get all location points in a given trail"
      parameters:
        - $ref: "#/components/parameters/TrailId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "None found"
        "500":
          description: "Server error"
  /trailfeatureslink:
    post:
      operationId: "linkhelper.newTrailFeature"
      tags:
        - Link entity
      summary: "Add new trail-feature link"
      requestBody:
        description: "Feature and trail to link"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailFeatureJSON"
              $ref: "#/components/schemas/TrailFeature"
      responses:
        "201":
          description: "Successfully added trail feature"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"

    delete:
      operationId: "linkhelper.deleteTrailFeature"
      tags:
        - Link entity
      summary: "Delete trail-feature link"
      requestBody:
        description: "Feature and trail to delete link of"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailFeatureJSON"
              $ref: "#/components/schemas/TrailFeature"
      responses:
        "201":
          description: "Successfully deleted trail feature"
        "400":
          description: "Bad request"
        "404":
          description: "No TrailFeature found"
        "500":
          description: "Server error"
  /trailfeaturelink/gettrails/{FeatureId}:
    get:
      operationId: "linkhelper.getTrailsByFeature"
      tags:
        - Link entity
      summary: "Get all trails with a given feature"
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "None found"
        "500":
          description: "Server error"
  /trailfeaturelink/getfeatures/{TrailId}:
    get:
      operationId: "linkhelper.getFeaturesByTrail"
      tags:
        - Link entity
      summary: "Get all features of a given trail"
      parameters:
        - $ref: "#/components/parameters/TrailId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "None found"
        "500":
          description: "Server error"
# Link entities
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #